days 16,17

+++ newDay.sh
@@ -6,7 +6,10 @@
 # Distributed under terms of the MIT license.
 #
 
-
+if [ -f "src/Day$1.hs" ]; then
+  nvim src/Day$1.hs 
+  exit
+fi
 cp template.hs src/Day$1.hs
 sed -i "s/10/$1/g" src/Day$1.hs
 sed -i "s/$(expr $1 - 1)/$1/g" Main.hs
